requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cors-headers==3.12.0
django-filter==22.1
django-rest-passwordreset==1.3.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.1.0
xlrd==2.0.1
